fix(core): update tab list when player levels up
